Refactor language button in footer as a dedicated I18nBtnChangeLocale component (reusability/flexibility)
